+++ runtime/service/src/main/java/org/apache/polaris/service/catalog/io/DefaultFileIOFactory.java
@@ -23,6 +23,7 @@ import io.smallrye.common.annotation.Identifier;
 import jakarta.annotation.Nonnull;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
+import java.time.Clock;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
@@ -39,7 +40,9 @@ import org.apache.polaris.core.persistence.PolarisResolvedPathWrapper;
 import org.apache.polaris.core.storage.AccessConfig;
 import org.apache.polaris.core.storage.PolarisCredentialVendor;
 import org.apache.polaris.core.storage.PolarisStorageActions;
+import org.apache.polaris.core.storage.StorageAccessProperty;
 import org.apache.polaris.core.storage.cache.StorageCredentialCache;
+import org.apache.polaris.service.storage.StorageConfiguration;
 
 /**
  * A default FileIO factory implementation for creating Iceberg {@link FileIO} instances with
@@ -53,13 +56,19 @@ import org.apache.polaris.core.storage.cache.StorageCredentialCache;
 @Identifier("default")
 public class DefaultFileIOFactory implements FileIOFactory {
 
+  private final StorageConfiguration storageConfiguration;
   private final StorageCredentialCache storageCredentialCache;
   private final MetaStoreManagerFactory metaStoreManagerFactory;
+  private final Clock clock;
 
   @Inject
   public DefaultFileIOFactory(
+      StorageConfiguration storageConfiguration,
       StorageCredentialCache storageCredentialCache,
-      MetaStoreManagerFactory metaStoreManagerFactory) {
+      MetaStoreManagerFactory metaStoreManagerFactory,
+      Clock clock) {
+    this.storageConfiguration = storageConfiguration;
+    this.clock = clock;
     this.storageCredentialCache = storageCredentialCache;
     this.metaStoreManagerFactory = metaStoreManagerFactory;
   }
@@ -82,7 +91,7 @@ public class DefaultFileIOFactory implements FileIOFactory {
     Optional<PolarisEntity> storageInfoEntity =
         FileIOUtil.findStorageInfoFromHierarchy(resolvedEntityPath);
     Optional<AccessConfig> accessConfig =
-        storageInfoEntity.map(
+        storageInfoEntity.flatMap(
             storageInfo ->
                 FileIOUtil.refreshAccessConfig(
                     callContext,
@@ -101,6 +110,32 @@ public class DefaultFileIOFactory implements FileIOFactory {
       properties.putAll(accessConfig.get().credentials());
       properties.putAll(accessConfig.get().extraProperties());
       properties.putAll(accessConfig.get().internalProperties());
+    } else {
+      // If no subscoped creds were produced, use system-wide AWS or GCP credentials if available.
+      if (storageConfiguration.awsAccessKey().isPresent()
+          && storageConfiguration.awsSecretKey().isPresent()) {
+        // If no subscoped creds, use system-wide AWS credentials if available
+        properties.put(
+            StorageAccessProperty.AWS_KEY_ID.getPropertyName(),
+            storageConfiguration.awsAccessKey().get());
+        properties.put(
+            StorageAccessProperty.AWS_SECRET_KEY.getPropertyName(),
+            storageConfiguration.awsSecretKey().get());
+      } else if (storageConfiguration.gcpAccessToken().isPresent()) {
+        properties.put(
+            StorageAccessProperty.GCS_ACCESS_TOKEN.getPropertyName(),
+            storageConfiguration.gcpAccessToken().get());
+        properties.put(
+            StorageAccessProperty.GCS_ACCESS_TOKEN_EXPIRES_AT.getPropertyName(),
+            String.valueOf(
+                clock
+                    .instant()
+                    .plus(
+                        storageConfiguration
+                            .gcpAccessTokenLifespan()
+                            .orElse(StorageConfiguration.DEFAULT_TOKEN_LIFESPAN))
+                    .toEpochMilli()));
+      }
     }
 
     return loadFileIOInternal(ioImplClassName, properties);