+++ CHANGELOG.md
@@ -29,6 +29,13 @@ request adding CHANGELOG notes for breaking (!) changes and possibly other secti
 
 ### Highlights
 
+- Support for S3 request signing has been added, allowing Polaris to work with S3-compatible object storage systems.
+  *Remote signing is currently experimental and not enabled by default*. In particular, RBAC checks are currently rather
+  coarse-grained. One new table privilege was introduced: `TABLE_REMOTE_SIGN`. To enable remote signing:
+  1. Set the system-wide property `REMOTE_SIGNING_ENABLED` or the catalog-level `polaris.request-signing.enabled` 
+     property to `true`.
+  2. Grant the `TABLE_REMOTE_SIGN` privilege to a catalog role.
+
 ### Upgrade notes
 
 ### Breaking changes